+++ js/View/GardenersCompendiumView.js
@@ -17,10 +17,8 @@
  * along with Cooker.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import _ from "../libad/Underscore.js";
 import UI from "../UI.js";
 
-const sectionId = "gardenersCompendiumSection";
 const expandedHeaderText = "Gardener's Compendium";
 const expandedHeaderTooltip = "Click to hide this Gardener's Compendium";
 const collapsedHeaderText = "G..";
@@ -29,34 +27,35 @@ const collapsedHeaderTooltip = "Click to open Gardener's Compendium";
 export default class GardenersCompendiumView {
   constructor(plantListView) {
     this._plantListView = plantListView;
+    this.sectionHeaderId = "gardenersCompendiumSection";
   }
 
   build(plantList) {
-    const header = UI.h2(expandedHeaderText);
-    header.attr("title", expandedHeaderTooltip);
-    header.addClass("nzc-gardeners-compendium__header");
-    const section = UI.section(header);
-    section.attr("id", sectionId);
+    const section = UI.section();
+    section.attr("id", this.sectionHeaderId);
     section.addClass("nzc-gardeners-compendium");
-    const list = this._plantListView.build(plantList).appendTo(section);
-    list.addClass("nzc-gardeners-compendium__list");
-    $(document).on(
-      "click"
-      , `#${sectionId} > h2`
-      , _.bind(this._toggleSection, this, header, list)
-    );
+
+    this._header = UI.h2(expandedHeaderText);
+    this._header.attr("title", expandedHeaderTooltip);
+    this._header.addClass("nzc-gardeners-compendium__header");
+    this._header.appendTo(section);
+
+    this._list = this._plantListView.build(plantList);
+    this._list.addClass("nzc-gardeners-compendium__list");
+    this._list.appendTo(section);
+
     return section;
   }
 
-  _toggleSection(header, list) {
-    if (list.is(":visible")) {
-      header.text(collapsedHeaderText);
-      header.attr("title", collapsedHeaderTooltip);
-      list.hide();
+  toggleSection() {
+    if (this._list.is(":visible")) {
+      this._header.text(collapsedHeaderText);
+      this._header.attr("title", collapsedHeaderTooltip);
+      this._list.hide();
     } else {
-      header.text(expandedHeaderText);
-      header.attr("title", expandedHeaderTooltip);
-      list.show();
+      this._header.text(expandedHeaderText);
+      this._header.attr("title", expandedHeaderTooltip);
+      this._list.show();
     }
   }
 }